requirements.txt
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
numpy==2.2.5
pandas==2.2.3
pdfminer.six==20250416
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
